build: Move bazel/utils.bzl into api/bazel (#26858)
Move bazel/utils.bzl into api/bazel

The api/... subtree depends on utils.bzl, but it is stored in //bazel,
not //api/bazel.  When the api sub-tree is exported as
"data-plane-api", the resulting repository is not buildable.

This attempts to fix that by changing the canonical location of
utils.bzl and updating the references to it.

Signed-off-by: Ryan Anderson <ryan@michonline.com>

Signed-off-by: Ryan Northey <ryan@synca.io>