bazel: fix gcc 12+ warning
I'm not sure why just moving this into a cc file seems to be enough, but
with this I can no longer reproduce the warning. If the warning does
still appear in this case, we can add `--per_file_copt` to disable it in
a single place, which isn't possible with a header which causes the
warning to happen in each file it's included by.

Fixes: https://github.com/envoyproxy/envoy/issues/23783

Signed-off-by: Keith Smiley <keithbsmiley@gmail.com>