ci: run Kotlin tests with signal_trace disabled
There are several Java tests that fail without the
`--define=signal_trace=disabled` flag. The flag is already on for Java
tests; this commit turns them on for Kotlin tests on CI as well.

Also, this solves one of the problems with flaky tests in
StatsFlushIntegrationTest.kt, as described in
https://github.com/envoyproxy/envoy/issues/24657, so we re-enable the
test that now passes. The other StatsFlushIntegrationTest test still
fails and will be debugged separately.

Signed-off-by: Ali Beyad <abeyad@google.com>