Add createPath to Filesystem abstraction. (#27052)
Add createPath to Filesystem abstraction
Additional Description: createPath support is a prerequisite of file_system_http_cache's create_cache_path support.

This is a revert of a revert. Now that mobile CI will run as part of the change, this should be safer to land when it passes than the two previous attempts. The revert is the first commit in this PR, so the changes that are new/fixes can be easily reviewed separately.

Compared to the previous commit that broke mobile CI, this change performs posix file system operations with C style operating system calls, thereby avoiding the prior problem of C++17 filesystem functions failing to link.
Signed-off-by: Raven Black <ravenblack@dropbox.com>