tls: Use BoringSSL APIs to bootstrap stat names (#27261)
ContextImpl predeclares stat names to avoid needing synchronization and
logging stats later. It does this by (1) iterating over
SSL_CTX_get_ciphers at SSL_CTX creation time and (2) hard-coding a lists
of known TLS 1.3 ciphers, TLS versions, NamedGroups, etc.

(1) implicitly assumes cipher suite configuration won't be applied
later. The OpenSSL API allows applications to configure ciphers on
either the SSL_CTX or the SSL. Envoy has an SslCtxCb which callers could
use to either reconfigure the SSL_CTX or install a callback that does
the latter. In those two cases, the final cipher list may include a
cipher that the initial one does not.

(2) breaks whenever BoringSSL adds a new feature. In principle, we could
update Envoy when updating BoringSSL, but this is an unresasonable
development overhead for just one of many BoringSSL consumers to impose.
Such costs are particularly high when considering needing to coordinate
updates to Envoy and BoringSSL across different repositories. (The lists
are already out of date, as we've replaced our old experimental
post-quantum algorithm, CECPQ2, with a new one X25519Kyber768Draft00.
That one is also experimental and will be replaced in the future too.)

To fix both of these,
https://boringssl-review.googlesource.com/c/boringssl/+/59667 adds APIs
in BoringSSL to query possible strings that various functions can
return. These lists will be a superset of those that any one application
may care about (e.g. we may have a deprecated cipher that Envoy no
longer needs, it may be a cipher only applicable for some modes, or an
cipher that's too experimental for general use) but this is fine
provided this is just used to initialize the table. In particular, they
are *not* intended to enumerate supported features.

Since Envoy has to support a range of BoringSSL versions, the new code
is gated behind BORINGSSL_API_VERSION. On older BoringSSLs, the it still
uses the old code, with all the issues unchanged.

Signed-off-by: David Benjamin <davidben@google.com>