Smtp proxy Starttls don't try to defer upstream communication until after downstream starttls for the sake of proxy protocol. New state machine is basically linear and supports any combination of upstream and downstream starttls.
I have dropped the proxy protocol support for the time
being. It seems like the downstream isn't actually connected as of
onNewConnection() so trying to send it there results in an empty proxy
protocol header. If you don't want upstream starttls, it may be
possible to wire up upstream_proxy_protocol as the upstream cluster
transport socket today and it will just work. In any event, the
current implementation of proxy protocol in Envoy never includes the
TLS info. XCLIENT may be a better solution to that.

Signed-off-by: John Bucy <bucy@gloop.org>