stats: add 'detailed' buckets-mode that captures all data from circlhist (#27094)
Commit Message: The current mechanisms for extracting histograms leave a lot of useful data inaccessible. This adds a new extraction mechanism to provide the full computed detail provided by the histogram library, as well as JSON and textual access to this data through the admin port.

Additional Description: This partially addresses #11259 but does not add the new detailed bucket mode to the metrics service gRPC API.

This PR sets the stage for admin visualization of histograms in #26472. The visualization could've occurred without this but it was not very rich; it's got more potential for revealing insight if the full bucket data can be rendered. Note that when aggregating histograms over multiple Envoy proxies it may be better to use the configurable fixed bucket sizes so that the histogram data can be combined across Envoy instances by the time-series database (e.g. Prometheus).

Risk Level: low: this is new functionality.
Testing: //test/...
Docs Changes: added doc for new bucket-mode.
Release Notes: required
Platform Specific Features: n/a

Signed-off-by: Joshua Marantz <jmarantz@google.com>