socket: user space io handle invokes drain trackers immediately upon write (#27499)
* Fix use-after-free in user space IO handle where a drain callback may refer to
a freed object (i.e. the connection on the other side of the internal listener,
which may already have died). The solution employed herein is to call and clear
the drain trackers on write(), since - as far as the handle being written to is
concerned - we don't care about the write buffer as soon as it has been written
to the peer's IO handle.

Signed-off-by: Eugene Chan <eugenechan@google.com>