formatter: add two tests NaN and Infinity (#27589)
#26464 removed JsonOrDie function but added ENVOY_BUG in the format call path.

The pre-release version of Protobuf includes this change:
Expect fail when serialize inf and nan for Value.number_value in json format. fixes #11259
protocolbuffers/protobuf@ca1cb1b

If we bump the Protobuf version, there will be an exception & crash using NaN/inf/-inf in protobuf.struct number_value when performing the serialization, while there is no check for this behavior right now.

Added 2 tests to catch this behavior change in the future.

Signed-off-by: Boteng Yao <boteng@google.com>
Signed-off-by: Ryan Eskin <ryan.eskin89@protonmail.com>