tests: Eliminate retry backoffs in the SDS integration tests (#27679)
When the xDS cluster is specified after the cluster that depends on it
(e.g. for SDS) in the Bootstrap config, the cluster isn't available when
making the xDS request, causing a backoff retry after [0,1s).  This adds
an delays to the tests.

This commit fixes the issue by specifying the xDS cluster first in
order, before the static cluster that depends on it, in the Bootstrap
config.

There are about 120 tests in this file, so we save on average about 60
seconds of test runtime everytime the tests in this file are executed.

Signed-off-by: Ali Beyad <abeyad@google.com>