build: adding an disable_exceptions (which does not yet work) (#27811)
The new build option simply compiles out all try/catch code, while leaving in the exceptions. This can not yet be successfully used as it turns up fno-exceptions which chokes on throw statements. This is by design as if we compiled out throw as well, config failures would be fatal instead of gracefully handled.

Risk Level: low
Testing: manual testing
Docs Changes: n/a
Release Notes: n/a
Part of #27412

Signed-off-by: Alyssa Wilk <alyssar@chromium.org>
Signed-off-by: Ryan Eskin <ryan.eskin89@protonmail.com>