mobile: Optimize the Swift-C++ interop generateBootstrap bridge
This change passes a reference, instead of a copy, to the
generateBootstrapPtr bridge function. Copying the EngineBuilder can be
quite expensive, since it can hold lots of large strings (certs, filter
configs, etc).

The Swift bridge's generateBootstrap method needs to be declared
`mutating` because the EngineBuilder reference is not constant (const&).
The method parameter cannot be const& because of an interop oddity
described in
https://github.com/apple/swift/blob/main/docs/CppInteroperability/InteropOddities.md.

Signed-off-by: Ali Beyad <abeyad@google.com>