api: generate Go protobufs to a module subdirectory
Per discussion with the go-control-plane maintainers, there is a desire
to move all the Envoy API protobufs into a new module that is distinct
from the go-control-plane xDS APIs. This is intended to eventually allow
users to update prodobuf definitions without requiring new releases
of go-control-plane.

This change updates the Go module path for API protobufs to
github.com/envoyproxy/go-control-plane/proto, which will be
an independently versioned Go module.

Note that this needs a corresponding CI machinery change in the
go-control-plane respository.

Signed-off-by: James Peach <jpeach@apache.org>