mobile: Adds authentication token support to the xDS EngineBuilder APIs
This change enables passing in an authentication token to the gRPC
headers for authentication with an xDS management server. The invoker
specifies the header key (e.g. `x-goog-api-key`) and value (the
authentication token itself).

This change also updates the GCP Traffic Director integration test to
use API keys (`x-goog-api-key`) for authentication instead of JWT
tokens.

Signed-off-by: Ali Beyad <abeyad@google.com>