[mobile]Fix CronvoyUrlRequest state transition and callback logic (#28052)
Commit Message:Fix CronvoyUrlRequest state transition and callback logic
Additional Description:
Suppose the application sends a Cronvoy request and keeps calling request.read() until the end is reached, and the response is structured in a way such that the endStream signal comes by itself without any data.

Before this change, the behavior at the end of the stream is:
EnvoyMobile calls OnData() and OnComplete() callbacks to Cronvoy. As a result, the application gets an OnReadCompleted() callback with no data read, but no OnSucceeded() because the stream hasn't reached its final state.
The application tries to read again, and OnSucceeded() will be triggered.

After this change, the behavior at the end of the stream is:
EnvoyMobile calls OnData() and OnComplete() callbacks to Cronvoy. In the OnData() callback, the stream state will be updated to reflect that the final read is done. It will not invoke OnReadComplete() because it shouldn't be called when no data is read.
When OnComplete() is called, the stream can successfully move to its final state and invokes OnSucceeded() callback.

Risk Level: Low
Testing: Unit tests
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: Android only

Signed-off-by: Renjie Tang <renjietang@google.com>