ci: Add test targets to the mobile compile options workflow (#28083)
All C++ test targets are added to the cc_test job in the workflow.

This can help us catch test build failures when the standard build
options are disabled as they would often be in a release build.

This update caught a bug in the sds_integration_test.cc, which
wasn't guarded with envoy_select_google_grpc in the BUILD
file, so that problem has been fixed in this commit as well.

Signed-off-by: Ali Beyad <abeyad@google.com>
Signed-off-by: asheryer <asheryer@amazon.com>