ci: Increase timeout in Swift integration tests
Some work was done previously in
https://github.com/envoyproxy/envoy/pull/24892 to increase the Swift
integration test timeout to 10 seconds to match the Kotlin test
counterparts and reduce test flakiness.

However, there are still some places where the timeout was low and it
caused flakey tests, see
https://github.com/envoyproxy/envoy/issues/28115.

This change increases the timeout to 10 seconds in the other Swift tests
as well.

Signed-off-by: Ali Beyad <abeyad@google.com>