signal: set altstack size a multiple of page size (#28338)
We use mprotect to protect address space for signal handling, and mprotect requires its 
arguments  to be aligned to a page boundary. Here, guard_size_ is the system page size
and altstack_size_ is a multiple of page size. altstack_size_ is initialized as the maximum
of 4 * guard_size_ and MINSIGSTKSZ while MINSIGSTKSZ may NOT be a multiple of page
size.

Risk Level: Low
Testing: N/A
Docs Changes: N/A
Release Notes: N/A
Platform Specific Features: N/A
Signed-off-by: Xie Zhihao <zhihao.xie@intel.com>