UHV: allow additional characters in URI path for backward compatibility (#26837)
Add flag that enables validation of the :path header character set compatible with legacy Envoy codecs. When this flag is false header validator checks the URL path in accordance with the https://datatracker.ietf.org/doc/html/rfc3986#section-3.3 RFC.

This option currently is true by default and can be overridden using the "envoy.uhv.allow_extended_ascii_in_path_for_http2" runtime value. Note that the default value will be changed to false in the future to make Envoy behavior standard compliant and consistent across all HTTP protocol versions.

In the compatibility mode header validator allows the following additional characters:

* HTTP/1 protocol: " < > [ ] ^ ` { } \ | #
* HTTP/2 and HTTP/3 protocols: all characters allowed for HTTP/1, space, TAB
* HTTP/2 protocol: also allows all extended ASCII (>= 0x80)

NOTE: the " < > [ ] ^ ` { } \ | characters are not explicitly prohibited by the RFC-3986, they are just not part of any defined set. # is only allowed as a fragment separator. Extended ASCII, space, TAB are prohibited.

In addition when this flag is true AND path normalization is enabled, Envoy will do the following:

1. all additionally allowed characters with the exception of the [] and \ characters are percent encoded in the path segment of the URL only. These characters in query or fragment will remain unencoded.
2. \ character is translated to / in path segment.

With the envoy.uhv.allow_non_compliant_characters_in_path set to false the header validator rejects requests with characters not allowed by the RFC in the :path header.

---------

Signed-off-by: Yan Avlasov <yavlasov@google.com>