envoy: Add a configurable pending queue
The current Envoy pending-streams queue is a simple FIFO. While, this is
useful in a lot of cases, it doesn't provide the best QoS when upstreams
get overloaded. This change includes the following:
- Making the pending-streams queue in conn-pool base more configurable
- Introducing an adaptive-LIFO queue that switches to a LIFO based on a
threshold of the queue size. This is useful because it can prioritize
recent requests (optimizing for p50).
- Introducing controlled-delay (aka CoDel). The code has been borrowed from
https://github.com/facebook/folly/blob/main/folly/executors/Codel.h
This basically times out older requests in the queue as opposed to sending
them to the upstream.

Signed-off-by: Bashar Al-Rawi <basharal@live.com>