Add watermarking and buffering to cache filter (#26841)
Commit Message: [http cache filter] Add watermarking and buffering to cache filter
Additional Description: Per #26821 the current implementation risks OOM if the network is fast, as the response body gets slammed into the InsertContext as fast as it possibly can. This change makes it so the cache filter behaves as the HttpCache API suggests it will; waiting until the previous block of data has been replied to before sending more to the cache. The filter buffers and triggers high-watermark if the buffered amount reaches the configured buffer limit. This change also makes it possible for the cache to abort an insert, as the 'ready' parameter to the callbacks is now used by the filter.
Risk Level: Low; WIP extension.
Testing: Added coverage.
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a

Signed-off-by: Raven Black <ravenblack@dropbox.com>