[balsa] Validate trailers even if not enabled.
This brings BalsaParser trailer validation (format and characters) in
alignment with http-parser behavior. The downside is that this change
makes BalsaParser enforce the header size limit (both on individual
lines and on total size of trailer block) even when trailers are not
enabled, whereas http-parser does not restrict size at all when trailers
are not enabled.

Signed-off-by: Bence Béky <bnc@google.com>