Fix an error in createSinkPtr interface
Originally, the interface was accepting a
Server::Configuration::CommonFactoryContext object, but this type of object is
exclusively available to HTTP filters and cannot be used from in transport
socket contexts. Instead, we just pull out the relevant fields from each of the
contexts and apply them that way. It was also considered that we could use
`std::variant` (or `absl:variant`), but variant types appear to not support
reference objects, so this approach was not pursued.

Signed-off-by: Ashish Banerjee <ashish.banerjee@solo.io>