UHV: Fix response status checks (#27125)
Change the code in H/2 and H/3 codecs that needs valid :status value before the header map was validated by UHV to not expect the :status to be valid. If :status is invalid the response is rejected when it is checked by UHV in the CodecClient.

For H/1 codec the status validation remains in the low level parser as it is a much larger change to move it to UHV. This reduces visibility into protocol errors due to malformed status (i.e. it is impossible to see the specific status value that caused the failure), however it will be mitigated once parser error reasons are plumbed into the reset reason strings that are logged in the access log.

---------

Signed-off-by: Yan Avlasov <yavlasov@google.com>