[ZK filter] Fix unit tests for ZK proxy filter create flags (#28888)
Previously, unit tests were not able to test the order of the CreateFlags enum. Therefore, unit tests could not detect the bug fixed in #28056. This is because unit tests converted CreateFlags into int32 then convert it back. The test will pass as long as certain CreateFlags enum exists. This PR improves unit tests by passing the int32 enum value and trying to compare the output with the enum name.

Signed-off-by: Zhewei Hu <zhu@pinterest.com>