Add capability for ActiveQuicListener to custom handle undeliverable packets (#28664)
Add capability for special UDP packet handling during hot restart

Adds a flexible argument for shutdownListeners, making it possible to pass an additional argument to QUIC listeners (or others); uses that to pass a callback to ActiveQuicListener and EnvoyQuicDispatcher. This is a no-op in production since the parameter is currently always nullptr, but is a first step towards allowing us to catch new connections on a QUIC listener and forward them to a new instance during hot restart. An important part of #19454
Risk Level: Low, essentially a no-op. Adds one nullptr check in the common path per nonroutable quic packet; zero touch to the common routable packet path.
Testing: Added unit test in both EnvoyQuicDispatcher and ActiveQuicListener.
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a

Signed-off-by: Raven Black <ravenblack@dropbox.com>