Fix ext_proc issue with BUFFERED_PARTIAL mode (#29330)
In BUFFERED_PARTIAL mode, after receiving trailer, if there is data pending to be sent in the chunk_queue_, those data need to be sent to the ext_proc server before sending the trailer there. Before sending, Envoy will first consolidate the chunks and send all the data in one gRPC.

The behavior is very similar to BUFFERED mode. There is code to handle such case for BUFFERED mode. But it is missing for BUFFERED_PARTIAL mode.
---------

Signed-off-by: Yanjun Xiang <yanjunxiang@google.com>