http1: `Connection: close` header lost in internal redirect (#29174)
Envoy should close the connection after finishing serving a HTTP/1.1 request with Connection: close header and also send back Connection: close in its response header. This change fix a bug where if the request is redirected via recreateStream(), Envoy will drop Connection: close on the floor and treat the redirected request as if it doesn't have that header. The RFC doesn't enforce the server to close the connection immediately, but it is good to keep the behavior in sync with/without internal redirect.

Signed-off-by: Dan Zhang <danzh@google.com>