TCP connection: detect and send RST (#28817)
This PR will add the support to detect RST from the peer and send RST by Envoy controlled by flag.

This PR added enableTcpRstDetectAndSend method to the connection object, and it is disabled by default. The filter or application can enable it by calling enableTcpRstDetectAndSend(true) to enable it. This is similar for the usage of enableHalfClose(). In the meantime, this feature will also be controlled by the runtime guard envoy_reloadable_features_detect_and_raise_rst_tcp_connection.

---------

Signed-off-by: Boteng Yao <boteng@google.com>