network: Fix incorrect int comparison with -Werror,-Wsign-compare
The `mobile-release` workflow has been failing:
https://github.com/envoyproxy/envoy/actions/runs/6162864940/job/16725227974#step:5:53.
This is due to a compiler error when the `-Werror` and `-Wsign-compare`
compilation options are turned on on certain platforms.

On certain platforms, socklen_t is an int instead of unsigned int,
making the comparison with an integer of size_t (unsigned int) cause a
compilation failure.

Instead of changing the `mobile-release` workflow to figure out the
right compiler options to make the error go away, we fix the problem in
the code to ensure we are comparing integers of the same type (both are
now compared as unsigned ints).

Signed-off-by: Ali Beyad <abeyad@google.com>