golang filter: need to provide a plugin configuration in the http filter if the configuration is required (#30020)
Currently, when the plugin requires a plugin configuration, using these per-route configuration examples will cause an error: proto: mismatched message type: got "xds.type.v3.TypedStruct", want "". Even the filter level's configuration is only used to define the filter order.

One solution is to provide an empty xds.TypedStruct{} as fallback. But then I thought about it, and an empty configuration is not necessarily a valid configuration. And technically, if there is no perRoute configuration, then the filter will still be executed. So it looks like the only way to go is to change the config file and add plugin configuration to it.

Commit Message: golang filter: need to provide a plugin configuration in the http filter if the configuration is required
Additional Description:
Risk Level:
Testing:
Docs Changes: Change the examples to ensure they always work
Release Notes:
Platform Specific Features:

Signed-off-by: spacewander <spacewanderlzx@gmail.com>