mobile: Fix flaky CronetUrlRequestTest::testFailures
Fixes #30199.

The test cases for `CANCEL_ASYNC_WITHOUT_PAUSE` in
`CronetUrlRequestTest::testFailures` will perform a cancellation
asynchronously, but the order whether `onCanceled` or
`onResponseStarted` is called is is not deterministic, which means
sometimes the response may be available if `onResponseStarted` is
called first before `onCanceled` and sometimes the response may not
be available if `onCanceled` is called first. This PR attemps to fix
the issue by ignoring the assertion that checks whether the response is
empty since it is not very critical for this specific test. The
assertion that checks for the error remains unchanged to keep the same
level of coverage.

Signed-off-by: Fredy Wijaya <fredyw@google.com>