test: Enable XdsTest to run without the test HTTP server
This required fixing a couple issues in XdsTestServer:

(1) Initialize Libevent before creating the Dispatcher. If we don't do
this, and we start the XdsTestServer without starting the
Http2TestServer, the LibeventScheduler fails the
Libevent::Global::initialized() assert.

(2) Add a Thread::SkipAsserts before calling the YAML to Bootstrap
parsing. The YAML utilities make some non-relevant asserts on being on
the main thread, which aren't applicable when calling the YAML utilities
from JNI and tests.

To remove the need for the Http2TestServer, we remove the endpoints from
the dynamic cluster returned in the test xDS response. This allows us to
test the essential xDS behavior without depending on an endpoint for
cluster initialization.

Signed-off-by: Ali Beyad <abeyad@google.com>