mobile: Fix Envoy initialization bug
Prior to this PR, the code did not check for the `EnvoyStatus` when starting
Envoy, which means if there was a failure when running Envoy, the users would
not know causing a bug that could be difficult to debug. This PR fixes the issue
by throwing an exception when envoy status is `EnvoyStatus.ENVOY_FAILURE`.

Signed-off-by: Fredy Wijaya <fredyw@google.com>