Add FULL_SCAN mode to least request load balancer
By default, the least request load balancer returns the host with the fewest
active requests from a set of N randomly selected hosts.

This introduces a new "full scan" selection method that returns the host with
the fewest number of active requests from all hosts. If multiple hosts are
tied for "least", a host is selected using an approximately random means that
is time and memory efficient. The selection probability among tied hosts may
not be equal, depending on:
- The host count
- How many hosts are tied
- How those tied hosts are distributed within the hosts array

Signed-off-by: Jared Kirschner <jkirschner@hashicorp.com>
Signed-off-by: Leonardo da Mata <ldamata@spotify.com>