mobile: Remove support for JWT tokens in xDS
We started off using JWT tokens for authentication to the xDS management
server (e.g. GCP's Traffic Director), but now that we have
authentication token support via XdsBuilder::setAuthenticationToken, we
remove JWT token support, as it is not recommended to use for this type
of authentication and it is directly supported only on GoogleGrpc, and
we are switching to EnvoyGrpc.

Signed-off-by: Ali Beyad <abeyad@google.com>