mobile: Implement JniHelper
This PR implements `JniHelper`, which is a small wrapper around JNI API
with memory-safety. Currently, `JniHelper` does not wrap every function
available in the JNI API. It only wraps functions that are currently used
by Envoy Mobile. However, it should be easy to wrap more functions if
needed in the future.

This PR also adds an ability to test JNI functions with `-Xcheck:jni`
enabled to conform with JNI best practices.

Signed-off-by: Fredy Wijaya <fredyw@google.com>