mobile: Improve the gRPC initial metadata headers API for xDS (#30691)
Initially, we introduced the setAuthenticationToken() method to the
XdsBuilder to enable setting authentication tokens (e.g. API keys) on
the xDS gRPC stream. However, authentication sometimes requires more
than one header to be added to the gRPC stream (e.g. `x-android-package`
and `x-android-cert` for API keys). Also, it's nice to have a general
API for adding gRPC metadata headers, not just authentication-specific
ones.

In this commit, we remove the setAuthenticationToken() API on
XdsBuilder, and replace it with a addInitialStreamMetadata() API.

Signed-off-by: Ali Beyad <abeyad@google.com>