xds: Don't rely on stats to keep track of warming clusters (#30703)
As discussed in https://github.com/envoyproxy/envoy/pull/30700,
ClusterManagerImpl relies on the `warming_cluster` gauge to determine
when DiscoveryRequests can be unpaused. However, if the stat is not
allowlisted (like it wasn't in Envoy Mobile, resulting in a no-op
implementation being used), there will be bugs in the resumption of the
Cluster type's DiscoveryRequests.

This commit fixes the issue by introducing a "last recorded warming
clusters count" variable in ClusterManagerImpl, which we use to record
the previous warming clusters size. This way, we don't need to rely on
the stat to do the record keeping and makes the DiscoveryRequest
resumption code more robust to different stat configurations.

We don't need additional tests, since this is an internal implementation
detail, and we already have tests in Envoy as well as a CDS integration
test introduced in https://github.com/envoyproxy/envoy/pull/30700 that
triggers the code path and validates correct behavior.

Signed-off-by: Ali Beyad <abeyad@google.com>