docs: Reorder sections in the upgrades.rst
The order of the sections in the upgrades.rst was messed up recently
since the new "CONNECT-UDP support" and "Tunneling UDP over HTTP" were
inserted in-between the "CONNECT support" and "Tunneling TCP over HTTP".
I moved the Tunneling TCP over HTTP section abvoe the CONNECT-UDP
support to fix the order.

Signed-off-by: Jeongseok Son <jeongseok.son@gmail.com>